requirements.txt
alembic==1.4.3
Babel==2.9.0
blinker==1.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dnspython==2.0.0
email-validator==1.1.2
Flask==1.1.2
Flask-Babel==2.0.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Moment==0.11.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
psycopg2==2.8.6
pyenchant==3.2.0
PyJWT==2.0.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2020.5
redis==3.5.3
requests==2.25.1
rq==1.7.0
six==1.15.0
SQLAlchemy==1.3.22
urllib3==1.26.2
Werkzeug==1.0.1
WTForms==2.3.3
